requirements_agent.txt
llama-index-core
llama-index-embeddings-openai
llama-index-llms-openai
llama-index-readers-file
llama-index-agent-openai
llama-index-vector-stores-pinecone
pandas
openai
pinecone
python-dotenv
